Move cleanNameForMaya to general utilities
This allows easier integration with our testing framework and avoids duplicate definitions for the maya separator
