+++ src/serlio/materials/ArnoldMaterialNode.cpp
@@ -57,7 +57,7 @@ void setUvTransformAttrs(MELScriptBuilder& sb, const std::wstring& uvSet, const 
 void createMapShader(MELScriptBuilder& sb, const std::string& tex, const MaterialTrafo& mapTrafo,
                      const std::wstring& shaderName, const std::wstring& uvSet, const bool raw, const bool alpha) {
 	std::filesystem::path texPath(tex);
-	const std::wstring nodeName = mu::cleanNameForMaya(texPath.stem().wstring());
+	const std::wstring nodeName = prtu::cleanNameForMaya(texPath.stem().wstring());
 	sb.setVar(MEL_VAR_MAP_NODE, MELStringLiteral(nodeName));
 
 	sb.setVar(MEL_VAR_MAP_FILE, MELStringLiteral(prtu::toUTF16FromOSNarrow(tex)));
