+++ src/serlio/materials/StingrayMaterialNode.cpp
@@ -21,7 +21,6 @@
 #include "materials/MaterialInfo.h"
 #include "materials/MaterialUtils.h"
 
-#include "utils/MayaUtilities.h"
 #include "utils/MELScriptBuilder.h"
 
 namespace {
@@ -41,7 +40,7 @@ void setTexture(MELScriptBuilder& sb, const std::wstring& target, const std::wst
                 const std::wstring& alphaTarget = {}) {
 	if (!tex.empty()) {
 		std::filesystem::path texPath(tex);
-		const std::wstring nodeName = mu::cleanNameForMaya(texPath.stem().wstring());
+		const std::wstring nodeName = prtu::cleanNameForMaya(texPath.stem().wstring());
 		sb.setVar(MEL_VAR_MAP_NODE, MELStringLiteral(nodeName));
 
 		sb.setVar(MEL_VAR_MAP_FILE, MELStringLiteral(tex));
