+++ src/serlio/modifiers/RuleAttributes.cpp
@@ -20,7 +20,6 @@
 #include "modifiers/RuleAttributes.h"
 
 #include "utils/LogHandler.h"
-#include "utils/MayaUtilities.h"
 #include "utils/Utilities.h"
 
 #include "prt/Annotation.h"
@@ -39,19 +38,19 @@ constexpr bool DBG = false;
 constexpr const wchar_t* PRT_ATTR_FULL_NAME_PREFIX = L"PRT_";
 
 std::wstring getFullName(const std::wstring& fqAttrName, std::map<std::wstring, int>& mayaNameDuplicateCountMap) {
-	const std::wstring fullName = PRT_ATTR_FULL_NAME_PREFIX + mu::cleanNameForMaya(fqAttrName);
+	const std::wstring fullName = PRT_ATTR_FULL_NAME_PREFIX + prtu::cleanNameForMaya(fqAttrName);
 	// make sure maya names are unique
 	return fullName + prtu::getDuplicateCountSuffix(fullName, mayaNameDuplicateCountMap);
 }
 
 std::wstring getBriefName(const std::wstring& fqAttrName, std::map<std::wstring, int>& mayaNameDuplicateCountMap) {
-	const std::wstring briefName = mu::cleanNameForMaya(prtu::removeStyle(fqAttrName));
+	const std::wstring briefName = prtu::cleanNameForMaya(prtu::removeStyle(fqAttrName));
 	// make sure maya names are unique
 	return briefName + prtu::getDuplicateCountSuffix(briefName, mayaNameDuplicateCountMap);
 }
 
 std::wstring getNiceName(const std::wstring& fqAttrName) {
-	return mu::cleanNameForMaya(prtu::removeImport(prtu::removeStyle(fqAttrName)));
+	return prtu::cleanNameForMaya(prtu::removeImport(prtu::removeStyle(fqAttrName)));
 }
 
 } // namespace
